BE: fix UserModel import
